refactored type imports and file location
