+++ lib/video-api/videos/queries/get-video-by-id.ts
@@ -1,5 +1,5 @@
 import 'server-only'
-import type { Video } from '../types/video-api.types'
+import type { Video } from '@/lib/types/video-api.types'
 
 export default async function getVideoById(id: string): Promise<Video> {
   try {
